% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trybox.R
\name{ggBoxPlot}
\alias{ggBoxPlot}
\title{ggBoxPlot}
\usage{
ggBoxPlot(listOflist, newOrder)
}
\arguments{
\item{listOflist}{A list of lists, possibly named}

\item{newOrder}{An index of the lists in the order you want them to be plot}
}
\details{
plot ggplot2 BoxPlot in the order you want.
}
\examples{
magic <- list(list(c(rnorm(20,2,1)), c(rnorm(8,8,1))), list(c(rnorm(20,3,1)), c(rnorm(8,7,1))), list(c(rnorm(20,4,1)), c(rnorm(8,6,1))))
names(magic) <- c(1:length(magic))
order= c("2","3", "1")
p <- ggBoxPlot(magic, order)

}
